requirements.txt
requests>=2.31.0
zstandard>=0.22.0
m3u8>=3.5.0
beautifulsoup4>=4.12.0
pycryptodome>=3.19.0
